feat: implement SEO updates and optimize asset handling across the application
