+++ src/website/src/app/(utility)/explore-data/mobile-app/page.tsx
@@ -1,124 +1,5 @@
-'use client';
-import { motion } from 'framer-motion';
-import Image from 'next/image';
-import Link from 'next/link';
+import MobileAppPage from '@/features/explore-data/MobileAppPage';
 
-import {
-  CLOUDINARY_IMAGES,
-  optimizeCloudinaryUrl,
-} from '@/services/external/cloudinary.service';
+const page = () => <MobileAppPage />;
 
-const images = {
-  hero: optimizeCloudinaryUrl(
-    'https://res.cloudinary.com/dbibjvyhm/image/upload/v1728132440/website/photos/explore/man-download-app_qg1pt2.png',
-    { width: 1200 },
-  ),
-  qrCode: optimizeCloudinaryUrl(
-    'https://res.cloudinary.com/dbibjvyhm/image/upload/v1770647202/website/photos/explore/frame_k2ncpc.png',
-    { width: 440 },
-  ),
-};
-
-const AppPromo = () => {
-  return (
-    <div className="flex flex-col lg:flex-row min-h-screen">
-      {/* Left Side: Image with responsiveness */}
-      <div className="hidden md:block w-full lg:w-1/2 relative h-64 lg:h-auto">
-        <Image
-          src={images.hero}
-          alt="User using AirQo app"
-          fill
-          className="object-cover w-full h-full"
-          loading="eager"
-        />
-      </div>
-
-      {/* Right Side: Promo Content with responsive padding */}
-      <div className="w-full lg:w-1/2 relative flex flex-col justify-center items-start text-left bg-white p-8 md:p-12 lg:p-24">
-        {/* Back button */}
-        <div className="relative mb-2 lg:absolute top-0 left-0 p-4">
-          <Link href="/explore-data" passHref>
-            <span className="text-2xl text-gray-500 hover:text-gray-700">
-              &larr;
-            </span>
-          </Link>
-        </div>
-
-        {/* AirQo logo with animation */}
-        <motion.div
-          initial={{ opacity: 0, y: -20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.5 }}
-          className="mb-8"
-        >
-          <Image
-            src={CLOUDINARY_IMAGES.logo}
-            alt="AirQo Logo"
-            width={80}
-            height={80}
-            loading="eager"
-          />
-        </motion.div>
-
-        {/* App promo text with animation */}
-        <motion.div
-          initial={{ opacity: 0, y: -20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ delay: 0.3, duration: 0.5 }}
-        >
-          <h1 className="text-3xl md:text-4xl font-bold text-gray-900 mb-4">
-            Get the AirQo app
-          </h1>
-          <p className="text-base md:text-lg text-gray-500 mb-8 md:mb-10">
-            Discover the quality of air you are breathing.
-          </p>
-        </motion.div>
-
-        {/* QR code and Download buttons with animation */}
-        <motion.div
-          className="flex flex-col md:flex-row items-center space-y-4 md:space-y-0 md:space-x-8"
-          initial={{ opacity: 0, y: -20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ delay: 0.5, duration: 0.5 }}
-        >
-          {/* QR Code - Larger size for better visibility */}
-          <div className="border border-gray-300 rounded-lg p-4 hidden lg:block">
-            <Image src={images.qrCode} alt="QR Code" width={220} height={220} />
-          </div>
-
-          {/* Download buttons */}
-          <div className="flex flex-col space-y-4">
-            <a
-              href="https://apps.apple.com/ug/app/airqo-air-quality/id1337573091"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="block"
-            >
-              <Image
-                src={CLOUDINARY_IMAGES.appleStoreBadge}
-                alt="Download on the App Store"
-                width={160}
-                height={50}
-              />
-            </a>
-            <a
-              href="https://play.google.com/store/apps/details?id=com.airqo.app"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="block"
-            >
-              <Image
-                src={CLOUDINARY_IMAGES.googlePlayBadge}
-                alt="Get it on Google Play"
-                width={160}
-                height={50}
-              />
-            </a>
-          </div>
-        </motion.div>
-      </div>
-    </div>
-  );
-};
-
-export default AppPromo;
+export default page;
