+++ src/website/src/components/NonBlockingFontLink.tsx
@@ -1,14 +1,17 @@
 'use client';
 
-const NonBlockingFontLink = ({ href }: { href: string }) => (
-  <link
-    rel="stylesheet"
-    href={href}
-    media="print"
-    onLoad={(event) => {
-      event.currentTarget.media = 'all';
-    }}
-  />
-);
+import { useEffect, useRef } from 'react';
+
+const NonBlockingFontLink = ({ href }: { href: string }) => {
+  const linkRef = useRef<HTMLLinkElement>(null);
+
+  useEffect(() => {
+    if (linkRef.current) {
+      linkRef.current.media = 'all';
+    }
+  }, []);
+
+  return <link ref={linkRef} rel="stylesheet" href={href} media="print" />;
+};
 
 export default NonBlockingFontLink;
